Add some extra tests

+++ tests/convert_test.rb
@@ -8,14 +8,41 @@ class ConvertTest < Test::Unit::TestCase
     assert_equal(nil, movie)
   end
 
-  def test_conversion
+  def test_mkv_conversion
+    movie = convert("tests/fixtures/SampleVideo.mp4",".mp4", ".mkv")
+    assert_not_equal(nil, movie)
+  end
+
+  def test_avi_conversion
     movie = convert("tests/fixtures/SampleVideo.mp4",".mp4", ".avi")
     assert_not_equal(nil, movie)
   end
 
+  def test_mpeg_conversion
+    movie = convert("tests/fixtures/SampleVideo.mp4",".mp4", ".mpeg")
+    assert_not_equal(nil, movie)
+  end
+
+  def test_ogg_conversion
+    movie = convert("tests/fixtures/SampleVideo.mp4",".mp4", ".ogg")
+    assert_not_equal(nil, movie)
+  end
+
+  def test_webm_conversion
+    movie = convert("tests/fixtures/SampleVideo.mp4",".mp4", ".webm")
+    assert_not_equal(nil, movie)
+  end
+
+def test_convert_folder
+  movie = convert("tests/fixtures", ".mp4", ".avi")
+  assert_not_equal(nil, movie)
+end
+
   def teardown
-    if File.exist?("SampleVideo.avi")
-      File.delete("SampleVideo.avi")
-    end
+      File.delete("SampleVideo.mkv") if File.exist?("SampleVideo.mkv")
+      File.delete("SampleVideo.avi") if File.exist?("SampleVideo.avi")
+      File.delete("SampleVideo.mpeg") if File.exist?("SampleVideo.mpeg")
+      File.delete("SampleVideo.ogg") if File.exist?("SampleVideo.ogg")
+      File.delete("SampleVideo.webm") if File.exist?("SampleVideo.webm")
   end
 end
